remove ReadOnlyTrie
