+++ src/neo/Cryptography/MPT/MPTDb.cs
@@ -4,20 +4,29 @@ using System;
 
 namespace Neo.Cryptography.MPT
 {
-    public class MPTDb : MPTReadOnlyDb
+    public class MPTDb
     {
-        private readonly ISnapshot store;
+        private readonly ISnapshot Store;
+        private readonly byte Prefix;
 
-        public MPTDb(ISnapshot store, byte prefix) : base(store, prefix)
+        public MPTDb(ISnapshot store, byte prefix)
         {
-            this.store = store;
+            this.Store = store;
+            this.Prefix = prefix;
+        }
+
+        public MPTNode Node(UInt256 hash)
+        {
+            if (hash is null) return null;
+            var data = Store.TryGet(Prefix, hash.ToArray());
+            return MPTNode.Decode(data);
         }
 
         public void Put(MPTNode node)
         {
             if (node is HashNode)
                 throw new InvalidOperationException("Means nothing to store HashNode");
-            store.Put(Prefix, node.Hash.ToArray(), node.Encode());
+            Store.Put(Prefix, node.Hash.ToArray(), node.Encode());
         }
     }
 }
