+++ src/neo/Cryptography/MPT/MPTTrie.Delete.cs
@@ -5,7 +5,7 @@ using static Neo.Helper;
 
 namespace Neo.Cryptography.MPT
 {
-    public partial class MPTTrie<TKey, TValue> : MPTReadOnlyTrie<TKey, TValue>
+    public partial class MPTTrie<TKey, TValue>
         where TKey : notnull, ISerializable, new()
         where TValue : class, ISerializable, new()
     {
