+++ src/neo/Cryptography/MPT/MPTTrie.Put.cs
@@ -3,7 +3,7 @@ using System;
 
 namespace Neo.Cryptography.MPT
 {
-    public partial class MPTTrie<TKey, TValue> : MPTReadOnlyTrie<TKey, TValue>
+    public partial class MPTTrie<TKey, TValue>
         where TKey : notnull, ISerializable, new()
         where TValue : class, ISerializable, new()
     {
